fix: use float64 whenever it is possible
