+++ .github/workflows/test.yml
@@ -1,5 +1,6 @@
 name: Test
-on: [push, pull_request]
+on: []
+#on: [push, pull_request]
 jobs:
   golangci:
     name: Golangci-lint
